+++ src/main.jsx
@@ -9,7 +9,7 @@ import store from './store.js'
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <Provider store={store}>
-    <Menu/>
+    <App/>
     </Provider>
   </React.StrictMode>,
 )
